feat: add status part

+++ work_muxixyz_app/api/status.py
@@ -1,6 +1,7 @@
 import os
 import time
 import requests
+import redis
 from flask import jsonify, request, current_app, url_for, Flask
 from . import api
 from .. import db
@@ -15,6 +16,8 @@ from ..mq import newfeed
 #KIND = ['Statu', 'Project', 'Doc', 'Comment', 'Team', 'User', 'File']
 num = 0
 page = 1
+redis_statu = redis.Redis(host='localhost', port=6379, decode_responses=True)
+
 
 @api.route('/status/new/', methods=['POST'], endpoint='newstatus')
 @login_required(1)
@@ -22,7 +25,7 @@ def newstatus(uid):
     content = request.get_json().get('content')
     title = request.get_json().get('title')
     time1 = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
-    statu = Statu(
+    statu =  Statu(
         content=content,
         title=title,
         time=time1,
@@ -32,7 +35,8 @@ def newstatus(uid):
     db.session.add(statu)
     db.session.commit()
     user = User.query.filter_by(id=uid).first()
-    action = 'create a new  status'
+    avatar_url = user.avatar
+    action = 'create '+ user.name + '\'s status'
     kind = 0
     sourceID = 0
     newfeed(
@@ -53,27 +57,33 @@ def getstatu(uid,sid):
     content = statu.content
     time = statu.time
     likeCount = statu.like
-    commentCount = statu.comment
+    iflike = 0
+    if statu.like is not 0:
+        likelen = redis_statu.llen(statu.id)
+        likeList = redis_statu.lrange(statu.id,0,likelen)
+        if uid in likeList:
+            iflike = 1
     user =  User.query.filter_by(id=uid).first()
     username = user.name
-    avatar = user.avatar
     comments = Comment.query.filter_by(statu_id=sid).all()
     commentList = []
     a_comment = {}
     for comment in comments:
         user_c = User.query.filter_by(id=comment.creator).first()
+        a_comment['cid'] = comment.id
         a_comment['username'] = user_c.name
         a_comment['avatar'] = user_c.avatar
         a_comment['time'] = comment.time
         a_comment['content'] =  comment.content
-        commentList.append(a_comment)
+        c_comment = a_comment.copy()
+        commentList.append(c_comment)
     response = jsonify({
+        "sid": sid,
         "title": title,
         "content": content,
-        "avatar": avatar,
         "time": time,
         "likeCount": likeCount,
-        "commentCount": commentCount,
+        "iflike": iflike,
         "userID": uid,
         "username": username,
         "commentList": commentList})
@@ -86,37 +96,41 @@ def getstatu(uid,sid):
 def editstatu(uid, sid):
     statu = Statu.query.filter_by(id=sid).first()
     if statu.user_id == uid:
-        content = request.get_json().get('content')
-        title = request.get_json().get('title')
-        time1 = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
-        statu = Statu.query.filter_by(id=sid).first()
-        statu.content = content
-        statu.title = title
-        statu.time = time1
-        db.session.add(statu)
-        db.session.commit()
-        user = User.query.filter_by(id=uid).first()
-        action = 'update '+ user.name + '\'s status'
-        kind = 0
-        sourceID = 0
-        newfeed(
-            uid,
-            action,
-            kind,
-            sourceID)
-        response = jsonify({"message":"statu edit successfully"})
-        response.status_code = 200
-        return response
+         content = request.get_json().get('content')
+         title = request.get_json().get('title')
+         time1 = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
+         statu = Statu.query.filter_by(id=sid).first()
+         statu.content = content
+         statu.title = title
+         statu.time = time1
+         db.session.add(statu)
+         db.session.commit()
+         user = User.query.filter_by(id=uid).first()
+         action = 'update '+ user.name + '\'s status'
+         kind = 0
+         sourceID = 0
+         newfeed(
+              uid,
+              action,
+              kind,
+              sourceID)
+         response = jsonify({"message":"statu edit successfully"})
+         response.status_code = 200
+         return response
+
 
 @api.route('/status/<int:sid>/', methods=['DELETE'], endpoint='deletestatu')
 @login_required(1)
 def deletestatu(uid,sid):
     if Statu.query.filter_by(id=sid).first() is not None:
-        Statu.query.filter_by(id=sid).delete()
-        response = jsonify({"message":"already delete the statu"})
+        statu = Statu.query.filter_by(id=sid).first()
+        if statu.user_id == uid:
+            Statu.query.filter_by(id=sid).delete()
+            response = jsonify({"message":"already delete the statu"})
+            response.status_code = 200
     else:
         response = jsonify({"message":"the statu has already been deleted"})
-    response.status_code = 200
+        response.status_code = 402
     return response
 
 
@@ -126,22 +140,34 @@ def statulist(uid, page):
     status = Statu.query.all()
     statuList = []
     a_statu = {}
+    iflike = 0
     for statu in status:
         global num
         num += 1
         if num > (page-1)*20 and num <= page*20:
+            if statu.like is not 0:
+                likelen = redis_statu.llen(statu.id)
+                likeList = redis_statu.lrange(statu.id,0,likelen)
+                if uid in likeList:
+                    iflike = 1
             user = User.query.filter_by(id=statu.user_id).first()
+            a_statu['sid'] = statu.id
             a_statu['username'] = user.name
             a_statu['time'] = statu.time
             a_statu['avatar'] = user.avatar
             a_statu['title'] = statu.title
             a_statu['content'] = statu.content
             a_statu['likeCount'] = statu.like
+            a_statu['iflike'] = iflike
             a_statu['commentCount'] = statu.comment
-            statuList.append(a_statu)
+            c_statu = a_statu.copy()
+            statuList.append(c_statu)
+        elif num > page * 20:
+            break
     response = jsonify({
         "statuList": statuList,
-        "page": page})
+        "page": page,
+        "count": len(status)})
     response.status_code = 200
     return response
 
@@ -152,22 +178,50 @@ def user_statulist(uid, userid, page):
     status = Statu.query.filter_by(user_id=userid).all()
     statuList = []
     a_statu = {}
+    iflike = 0
     for statu in status:
         global num
         num += 1
         if num > (page-1)*20 and num <= page*20:
+            if statu.like is not 0:
+                likelen = redis_statu.llen(statu.id)
+                likeList = redis_statu.lrange(statu.id,0,likelen)
+                if uid in likeList:
+                    iflike = 1
+            a_statu['sid'] = statu.id
             a_statu['time'] = statu.time
             a_statu['content'] = statu.content
             a_statu['likeCount'] = statu.like
+            a_statu['iflike'] = iflike
             a_statu['commentCount'] = statu.comment
-            statuList.append(a_statu)
+            c_statu = a_statu.copy()
+            statuList.append(c_statu)
+        elif num > page * 20:
+            break
     response = jsonify({
         "statuList": statuList,
-        "page": page})
+        "page": page,
+        "count": len(status)})
     response.status_code = 200
     return response
 
 
+@api.route('/status/<int:sid>/like/', methods=['PUT'], endpoint='like')
+@login_required(1)
+def like(uid, sid):
+    iflike = request.get_json().get("iflike")
+    statu = Statu.query.filter_by(id=sid).first()
+    if iflike == 1:
+        redis_statu.rpush(sid, uid)
+        statu.like += 1
+    if iflike == 0:
+        redis_statu.lrem(sid, uid, 0)
+        statu.like -= 1
+    db.session.add(statu)
+    db.session.commit() 
+    response = jsonify({"message":"change like number"})
+    response.status_code = 200
+    return response 
 
 @api.route('/status/<int:sid>/comments/', methods=['POST'], endpoint='newcomments')
 @login_required(1)
@@ -187,11 +241,9 @@ def newcomments(uid, sid):
     avatar_url = user.avatar
     action = 'comment '+ user.name + '\'s status'
     kind = 3
-    sourceID = db.session.query(func.max(Comment.id)).one()
-    sourceID = sourceID[0]
+    sourceID = comment.id
     newfeed(
         uid,
-        avatar_url,
         action,
         kind,
         sourceID)
@@ -200,41 +252,43 @@ def newcomments(uid, sid):
     return response
 
 
-@api.route('/status/<int:sid>/comment/<int:cid>/', methods=['GET'], endpoint='getcomment')
-@login_required(1)
-def getcomment(uid, sid, cid):
-    comment = Comment.query.filter_by(id=cid).first()
-    if comment is not None:
-        user = User.query.filter_by(id=comment.creator).first()
-        username = user.name
-        time1 = comment.time
-        avatar = user.avatar
-        content = comment.content
-        response = jsonify({
-            "username": username,
-            "avatar": avatar,
-            "time": time1,
-            "content": content})
-        response.status_code = 200
-    else:
-        response = jsonify({"message": "can't find comment"})
-        response.status_code = 402
-    return response
+#@api.route('/status/<int:sid>/comment/<int:cid>/', methods=['GET'], endpoint='getcomment')
+#@login_required(1)
+#def getcomment(uid, sid, cid):
+#    comment = Comment.query.filter_by(id=cid).first()
+#    if comment is not None:
+#        user = User.query.filter_by(id=comment.creator).first()
+#        username = user.name
+#        time1 = comment.time
+#        avatar = user.avatar
+#        content = comment.content
+#        response = jsonify({
+#            "username": username,
+#            "avatar": avatar,
+#            "time": time1,
+#            "content": content})
+#        response.status_code = 200
+#    else:
+#        response = jsonify({"message": "can't find comment"})
+#        response.status_code = 402
+#    return response
 
 
 @api.route('/status/<int:sid>/comment/<int:cid>/', methods=['DELETE'], endpoint='deletecomment')
 @login_required(1)
 def deletecomment(uid, sid, cid):
     if Comment.query.filter_by(id=cid).first() is not None:
-        Comment.query.filter_by(id=cid).delete()
-        response = jsonify({"message":"ok"})
-        response.status_code = 200
+        comment = Comment.query.filter_by(id=cid).first()
+        if comment.creator == uid:
+            Comment.query.filter_by(id=cid).delete()
+            response = jsonify({"message":"ok"})
+            response.status_code = 200
     else:
         response = jsonify({"message":"can't find"})
         response.status_code = 402 
     return response
 
-
+'''
 @api.route('/status/<int:sid>/comments/', methods=['GET'], endpoint='getcommentlist')
 @login_required(1)
 def getcommentlist(uid, sid):
@@ -244,11 +298,13 @@ def getcommentlist(uid, sid):
         commentlist = []
         for comment in comments:
             user = User.query.filter_by(id=comment.creator).first()
+            a_comment['cid'] = comment.id
             a_comment['username'] = user.name
             a_comment['avatar'] = user.avatar
             a_comment['time'] = comment.time
             a_comment['comment'] = comment.content
-            commentlist.append(a_comment)
+            c_comment = a_comment.copy()
+            commentlist.append(c_comment)
         response = jsonify({
                 "commentList": commentlist})
         response.status_code = 200
@@ -256,5 +312,5 @@ def getcommentlist(uid, sid):
         response = jsonify({"message": "can't find"})
         response.status_code = 402
     return response
-
+'''
